+++ src/components/pages/main_page/ai_image_tool/AIImageGenerator.svelte
