+++ src/components/pages/main_page/ai_image_tool/ai_sample_data.ts
