+++ src/components/pages/main_page/display/ai_sarga_translate/SargaAITranslate.svelte
