+++ src/components/pages/main_page/display/ai_sarga_translate/translation_prompts.yaml
